Refactor: Extract decrytor to internal package
