+++ internal/decryptor/testdata/age.txt
@@ -1,0 +1,3 @@
+# created: 2021-03-31T09:51:59+02:00
+# public key: age1l44xcng8dqj32nlv6d930qvvrny05hglzcv9qpc7kxjc6902ma4qufys29
+AGE-SECRET-KEY-1RH87A5Z54ZGUR9S0AS3R6WHFSEFNPLYMAKF5Z2CEU7R06VRJ0A3Q7242AM
