+++ internal/decryptor/testdata/pgp.asc
@@ -1,0 +1,117 @@
+-----BEGIN PGP PRIVATE KEY BLOCK-----
+
+lQcXBGBkJjIBEAC/ZvDZ8mR7m5AeeLCb0hgUn0TQBOL9Tn6FLOMkQDzRpytUkG5G
+lkEU/2FN2IjSZR1hulW8n8Wh5yDo6jNzneg2tjxBbBaJo4Mwd2YIs9xyzCVdvv1k
+UhHFCYUmhZeLizoRvTuMe2gl/OSrPsYy0mhbLMwr2v28PaaEljRX/2qJlB/u6fl2
+8xzGvBnV53LOPWB+j1WYfPQAKXVMShf8Ul+ZaT/vGRZ1ZUHJGp1Gs3vhyDQ2Zgxt
+UZJ/IpGnzvZAeTlBFUtfl2cma+bCcJbAvqoyDCHmZDwryINHtH20kN7cDRfrj08q
+hfqWqm4NrIIkY3llxJMApnkaYORWRGX62NdQbCgbX46ooJNXOBAq5fdBYyTWGwEE
+S1qfnYlz6zgyG+jZH1bTyYowm0gMzOm4Tv+UPKPgN09KV29155xKFzDCK3E+Bul5
+W4X1++gMBa33Up85GnHhAvVWxMbEXxVqzUs09I4CTW9f2Zx+WNhRunF1JcsjAptq
+kyNPk3/uAEsDIbhTClu2gfGUcUH8GGYkcMJgkHgB/ypjxV0Wtws17Y6sf477siqx
+IYTyar1qE/ab7QjDVULrcvGXyCv7OkHDUSYvugh0NmEyoaLVOuwKscZgGiHD8DtT
+NKRFWAaGy/gxpDXl4KM2DquDsFhaEykdu25YNOj2mYlNUc+qEx4bW2RWlQARAQAB
+AA/4hZqXK68lD7I5pgWTBnNBnV8MnbDXdEFKko3Ippcqj6UDWIcNwFuwmz8Nyrox
+zYGdy0ClnkaPTqP7lmogSpQqlb7a06ZMFSDwZmF018f2eOoTZrYaWhBskoK5aQCt
+5dgILdbL0j4K1EnixdX56NkmIsYQlTL6nYlZZZ61DrfuKQj1EdBMyYWonIHyLYG1
+JMRlwJjntbkvwtVdRw0j7JI6EVM27mHuCeCMseAFxSS5y8MpzNsD0apsgF4kAMNw
+5agvzAjVsfwVLkWUkxOBUV0TZI6QgH13R2wJDb0HZ253HBGocTC1mwUgRbf/Gkrl
+Y4ceJHpsLn8XtmPr+WmJrjbMizGne938QaqV9LQhqo12UEC4SY02xH6jnjRYhe2l
+zOuhl7QW494i3xuESiG2Zw83iXGBFgqyprhcbMrmmOhYqNgZAhH94mgMrJw3rXVu
+YQCE+wCjt81qkdNzlgPJELh/x5NZCNzhY8UqQgkIBeVkM/bH0bVSJPDjUvnKfg8i
+8I/n0W1RWyBixnv1ukEgTM4EOWpzB9zf94V+UImZ1s3Az13JQRPm6VedmSgwkxmV
+1CbfUkuTuySeafolc8uFqdBim4NcL0Ebc//saySeEoposD7BZ3WaSEx3NcVqhH4t
+FU2V5zNpHOakHo69vFYNY7HY9Mm6jnua+rhByOjCiLuKEQgAxeOQ4jfXvS0i3qRy
+ebF4ZmgWZ3qsCL0fUHa7tKdQ6OaRVNDLlZmYdbyQKSzGaXX+u6QJE6MUCmDoz8sS
+cz2Uo4s8JQl0s7XLXS7tLCMnRc5xDk+axntUig4BeSTdSmemA7Wawdx84ljvgs+w
+FSkUFk4Q86Ig/UPMmZzNtxg8Xp0L8cWwNfffo2W3LETI8OpU7l5U2FUNTnmgrbq0
+KTVAUsnBvn9lnLaZ3Vdha85jgSjUh1eJ9gVYrjo8zOnyJeHkyYgkzf+7vvlvlw9u
+wYjzSqWdby9zZD1/QGvGJv7iLBs9GJgqA+zFr0ez/vbi3mvU1hVRwytD6wAFaBYF
+ABZ8KwgA95u5WesYSJseUAikUc1oC9ko7IsVtmfEheU70YHPQtC0ScedHr8YB97X
+S9j2678I5bim4cACKY7GvinpjkZ8+6+lwZeWaKyRwQbvI4O3PfgNfxXZIFUHLfb4
+xLl//i6rOA8Z807ksviOhAdIKbg+3nAOBs+FPufLRGIn1IY8EPzhxar1RNDeRM+0
+n64F18+kO+vcNo98JLR0MWonXg2/ugrVf845vp7SlNT1844rZ6wMUMNuP8ldw7gx
+DTJHl5uChd6xvhWfi1WCB6h+fmjGKObuSsPQslhJFaQzULCwbozwBGU3kqB+2i9p
+kPfrmb0VQA+eZ3Wpc8P+JmL4ogdYPwf/Xv3EBs0+Eu8TjC6qvVWvwNwsFHsSJzJI
+OsRkwkSpcdaxsaIZTAKKWglmeoWOP9l/UFXjev0Cg3heolslt3LtS4m5VVwhM/Yq
+nDVm2357vcjh8UsvaUEiOCyPOKg+4sVerpJ/K7eL66cHbtaNxibUTkzkYkjq+GPB
+DxUhpwoL9SDaxFqb3x0kRwJx/A7uhII8L0SaUBtvTYJ5/vF97GKsj1skD3b4dSGh
+sCsInWqxvo5q2dU5NDlqtcyZJw/UyRSEkaLTQuSfYP9PcJTkI0AZcD2HEmyQg2Id
+F939QECktRQB1cTrlgytYGpYes3wXMRu7vCWEfX/foY2hWi15sZ1fIMItBl0ZXN0
+LmZsdXhjZC5pbyAodGVzdCBrZXkpiQJOBBMBCAA4FiEENcGmTNf8CrbrZnVrJEVG
+PDI07OEFAmBkJjICGy8FCwkIBwIGFQoJCAsCBBYCAwECHgECF4AACgkQJEVGPDI0
+7OHlUBAArG6/DLAaQasE1DgzPjncUn+1qKs6E3wSuI4YMCJlqm+y6TB0jJq9MlNf
+n3Bh919ZZHjWo0FSFjfguE7ySBVkkJd3GxhvIPthfQUOUMdYON3T4wBjvLOGOo4X
+SxmLw3rUy9D0yC6816m4qgXhZFSTj7bg1wRbUO43B7FMRsbNxgECDa4hEsueMFX4
+S26iIiMkdvSKTEmMSS9Qo5B5C0CnMpkjEgkqd24VCGRLcByM7N717l0Xfi4q0LMT
+UiE34/sXGsehqfh4JnU90RXNYHiIg6ASbsUT/sNUD76ucl8e7t9SbfWD2OYfbQGQ
+bJAOrsU8SnlWRGKWSvUK8Hgn6RePU6pvt2X6v3FhVol764KiegtJWimh5/x4BYxp
+x2rhXKNUGtdqTZ3r15NnMSSQHdnHDJO0T0/XioX89ghNqU+Li8nMyxAAtZATxE4x
+ATTwFQmbMDtYoU9Df+h2ej0uep/NWpOyrPPC1OqbZScEMdzRHteNExODs7/ehqlO
+LRjw96fyGwG1GgF/5qusZsLZg+7aZzG5itEOcMzLPXnQUq08+kWGmL7Ob2HbX3Fq
+eIuRz5iDKYhBMc/nGkXgZMFP4KO6Om29BqgYzjPY+NtJNnOxMH0PTwfAjau2PycS
+CmWrDU14MmXK71YxEPcdOKAz+ucGo3UXHoGQRfRpX6JysVRBN16dBxgEYGQmMgEQ
+ALjQCFtTYVLJKTkcKz991LjfbMZKt5ajIn2qDT2MSShXFi54DGb5LNufkOmHw2A3
+LM811xN//1cbvXa5/6p3LYiqmgVdK/b8ZudkGqhIH5oDGZfIRqY+9Bv63PfZDLu+
+SR9KmyOzmUzxBq0A+YnDC59KHa+9JtxeKgGrcXPF2d8oTpS3QExUsoH6EQJlEKfg
+d2F3yPiVvBwKcFpfBXrd+OO+iyZucOgT2/6fSpF1/pQHOXQfN5RPZ7bzsVU7Y2JD
+U3z0gmaMg5gF1eyeZtLi9cx8w23bC3f2SfmGhtdpsV07MmL+p6ql7l/xd1ItxCxv
+NJ9eAbxSOqUGfUM5SRqCzhh0NPtrCHzTd3MHLbBJDSPVL06btiuV+FKsOqvmW8ze
+WogPc1UszYN15Fz7G383AO7dadwczmfvGg/icHzWWWlw4ezT9YB0XrWZkLI325te
+wheiFIoWueQlUBLzu9p8m4W39KDguB2FCkrTB0kdZDR7KJslumRd7ywS9vv7jUsq
+zNkWViS1KRV1t/adr6aiIN3d8uLj5D+G/yfuUe04k4X39zUnP2nsqxQU+z9L4ugH
+sHgZEd7qfYP1x/o9u8PdzOLNKqk8RwI3ef2hSJkbUzV1QhRGBAkLKx1pQcUmECNV
+Ojt1TYZ6pzw3Y8aya3gIfG9ferR0RDwp0XOqou/PoDDRABEBAAEAD/45237SQG82
+gJ10MJplfRosal3eUiPZd+bjWmpcu19nzqO2FoUjc5yUgQbX5VXH+F4B3MQhaa/h
+cGDHG4GrtzW9FFO1nC6YWvpINHGX/0RNbe0ENIHRp+G6JLYqrXusBu1Ta4LYEXDL
+wqUpYiW8+B/aVVSn4NqaoKwb67VP3j/P+glny1rIyrc7s7LBgixNshsNm3eCQzYp
+vSbhU+AeUAGMIunjNS4JqbX0p5+rnDn/06cr8F3xA07f2rjIixgY+iTX+KAeRJ0p
+6aHIOurL2JvE0cZIzQtxGdyczn3GUGNjYZXrofeRl/LhfHGze3pvwrzBq6d1cnTD
+lx19/yAguZSVUIF+Qr4bugTMCLKeh0a2z4/UOxg/PEdgR6aDLzYbIpN73tyT6Q7A
+/rfX5GwBMCg7vE3eNNeSdxKre1kp6VRaOo0ylHybS1Xu+TI9e6JVs70N1WdbOyJ/
+UTXKXibHk/Xk4W/6DzBpJAiop0dBgmXUwO7w3FHa0QNmIlNYFz5+qzm4zIQjR7jE
+u/8z9JbyRwAB0QXVPO91MoZwWy5AHDP55agMyX2pl1wr9PHHYoSlsuLkGNz5lhLN
+Y2qsApNCKmYGzutc3eIypWDKujGkwIz+b5M7miFeNe/fo5htNfb+EDwsHmbPTqTg
+AVLMIMRJZLVYWmzpv8VX2C7YGS3loP+CBQgAxoacQap47YhVS27CvUMSM/a7e7GJ
+2GjJNo2sRpOY2Rm4IaRb6Snl8F+C5UBCtL6tb+jwKHMeYnlc4/jPPTV0TBX5U0av
+QmrYyACY8js6r3t8QFlX5/oZ+B0LahW87dEQIYc8cGrUbXn5a9pDAxZa4RcAYgNC
+Zx7/fIq5W9laYUXJHZ238Eln6GMIZIvHtYZlfprxfTTkL2AKwiSyStgU2vMgEqQq
+7btAbCmyZ4idVod09m9DWpUB2AMdIc0/clvVobLsEktHkxyw56LQlFpqJXEU89Gn
+3N25nIPK74TOwWxdIpfNyQszrKg5eMDvoSowI9tFwRwqU1Y/IxX721eZEwgA7lEX
+2ZghHmm4D5YPuR3txB/79iza4XNhfnmX7Wz2tdORBph5mXNm2hfaizIYyKnx0JrB
+knx2Bf6JBQXH2DkjKWumBYct1BgkplNpkOSM4QBjoHP6sHF5wCHtmQghF5Z677Fd
+AzSjb46pkczM4qzjj9DsidxmsQyM/erdIyh3+LMx4d/+UgFy358SfDuN/wEMt3BS
+F4HAAreL2jRmrnIJJK2CGg2wNBsxybZiANadTHKBMre9Jp5ciA0dhvKQLYBIqbmY
+/SBE2QBak0mJw2Dea47TwfuTCXfBj+BRWcpg7V1K23nlS4d2YOXulG85uIcgLetR
+cbLZSI3XzyimKJmPCwf/b198e7IOczGGcb1H/5gVkBEgRx7MdvKQmix6SQ75G288
+1yLTYEMqbQz6gW31ml/N/E62XGLQUSLvYzar4g8KPpKbjv8SzkOpZ6mYzQou/Ef7
+VAk9kGZjG9QngR9YUZ190zGH93KtCJJ7/HnhL1kAE9IksValY3Aw10h0n3wouA0F
+tR7TtLxKYcdsbSRpgCv6LG21EMW6+Np3C7356+wK4uUFydmUHK7qpQVQQl8r3HWb
+PzPeHcMK3F0V/r+uQIYNfURf3EU/fjhvj91yPxVGd26RRuwpLROu0CkmRE0WIES4
+ZM+16hk2VCwPU9VoZO8qcXcv3wcBmBp+DNwV5b4J2H7siQRsBBgBCAAgFiEENcGm
+TNf8CrbrZnVrJEVGPDI07OEFAmBkJjICGy4CQAkQJEVGPDI07OHBdCAEGQEIAB0W
+IQQVlMW3llrIrZg7yLDdEjiYoWgIywUCYGQmMgAKCRDdEjiYoWgIy8rjEACJUdH4
+FT1oEg5tGKW5NaFroSEX08DNGkWU45iUNclirewBQ3R+QhWgJTda0Uts8x3Vh2xX
+RzYZ8CpR23lUeq8pct1qdLNxm2SRep/g6uv0cow3WtY2SfkKlKud52hB3qrfkfxm
+A3laqRRbzsWLOIUHn5gYc+dG3maVUep1wcEI5AYuf0UZlHORNWWkOW4lscFkciF0
+A5MBATiYgcaPvcvPCFi15O5CRNC2sB0LQFN3y+IHD7SlVPWXX00p5HnC89tpgosK
+JKGXmX6w8/DW5t1rItoM+x2gDcmQj5UYrHL7xRFGghoNOJ48YwivyDfJmFchlg9j
+GZTCHUFPiHbmPXcuZPO19EyE63sFhClNpfUGX+/PxKWh2eAgSzTqP6gQst+ElPOE
+d8zz4Gwpb2mVsHKO5gxr40hE7CWPcJlpr+lBCzjXbEwnZRmzCUXMu+d5DxZg1gmC
+E8usVYDMT761r83+YfSXDEtCz551UwIHen7D6s4gompPSANBKZ0efYL20cE2PDxr
+GMWVHIRO+eNwm5dN8kW0or34Abk/pogIAdggkGwiwQVhoHTeXRAVi9Ex0ArjaKuh
++ij5mp2755pypEQlNsq6XMrOPUh5SGgqnZlz+vEqhKy+3tVezzQmOwXydhcNVUJR
+rEUvVVXCXTvPTwAC1D4O/MfhrZcomxgpPJm+VsxVEACc0ZhOKpeZPN4QYCHrvXLg
+OI82gWAF6Uq+Rzr15ZEYbn8D1GvFhDGPTF0eO1zVE/AurwmL8dXcw6Ps11URLchl
+RxQRByasdpA8WjCY4PsZPXMF0buqKEUN9MS9hJHBZsahbfYiTsHkXNhBaeVnIC0E
+7h3NXVigNingH2QHpQlNz3cSrgnxTUIwQFpOBXCCllh5zbZnXVtQsPjy76uRbRIs
+Xq8GFMh487lSk0AmMIIgjZznwHVO08QuEJPSv7POOHnjRuesaDTLRVcy7DFxpVcK
+mWMK/0/9vFmaD+8pifAr5j5A6sppocswZHneYHv8d1kPJGqw82DJp3veZeGB06Ys
+q2GQ0lBSEo2/g6jZZ3mxEqWB4IbVYgW3Tx7KMsRwLehm/lW5HVW4p7gMNjVa9sss
+Oi06LkwmDQ4UgFR09qsSVrBBWBRFuRq/z37KQ8UfZu00rhMZIhCYLMoIA/HhVvKj
+L3C3qtyQvqIk9U8QNW785EG2dvmadtdrqev4lz0rItCUXZey2sdaA0L/lY9CDkLe
+1e8DX4W1/oCCjTCrTPRcFwUVz7vIKe90xgclO8UWCJCCZ/faUJzevhQDt+dCBIc4
+srtsZ46p3UpAFT9qa+XZryYpTCD1m2Hs1vJx1f8fkvIhPY+8LBaWvMs8gBTIHAOu
+AU95IsCew/EYLvVp9luWxQ==
+=7gCf
+-----END PGP PRIVATE KEY BLOCK-----
